Removed Ruby script to test lobby features; The script is now replaced by a python version using autobahn|py. Future scripts can be written in other languages as many WAMP implementations for various script languages seem to be available.
A default crossbar config file is now copied into the build directory which will also start the STC_Server.
It is also configured to listen for connections on ports 8000 and 8080.
When debugging an executable this is not so usable and it is easier to start crossbar.io without the Server.